better structure for menu networking
