+++ Common/src/main/java/com/almostreliable/almostlib/menu/network/synchronizer/MenuSyncPacket.java
@@ -1,4 +1,4 @@
-package com.almostreliable.almostlib.menu.synchronizer;
+package com.almostreliable.almostlib.menu.network.synchronizer;
 
 import com.almostreliable.almostlib.AlmostLib;
 import com.almostreliable.almostlib.menu.SynchronizedContainerMenu;
