+++ Common/src/main/java/com/almostreliable/almostlib/menu/network/synchronizer/handler/BlockPosDataHandler.java
@@ -1,6 +1,6 @@
-package com.almostreliable.almostlib.menu.synchronizer.handler;
+package com.almostreliable.almostlib.menu.network.synchronizer.handler;
 
-import com.almostreliable.almostlib.menu.synchronizer.AbstractDataHandler;
+import com.almostreliable.almostlib.menu.network.synchronizer.AbstractDataHandler;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
