+++ Common/src/main/java/com/almostreliable/almostlib/menu/network/synchronizer/handler/package-info.java
@@ -1,0 +1,6 @@
+@ParametersAreNonnullByDefault @MethodsReturnNonnullByDefault
+package com.almostreliable.almostlib.menu.network.synchronizer.handler;
+
+import net.minecraft.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
